# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface BossKeyOverview {
  /**
     * @minimum 0
     * @maximum 100
     */
  readinessScore: number;
  /** @minimum 0 */
  missingConnections: number;
  /** @minimum 0 */
  expiringTokens: number;
  /** @minimum 0 */
  failedValidations: number;
  /** @minimum 0 */
  pausedWorkflows: number;
  /** @minimum 0 */
  inventoryConflicts: number;
  /** @minimum 0 */
  securityAlerts: number;
  discoveredAt: string;
}

export interface DiscoveryInput {
  projectPath?: string;
  repository?: string;
  branch?: string;
}

export interface DiscoveryResult {
  projectName: string;
  repository: string;
  branch: string;
  providersFound: number;
  environmentIssues: number;
  backupCreated: boolean;
  manifestPath: string;
  completedAt: string;
}

export interface Provider {
  slug: string;
  name: string;
  category: string;
  status: string;
  accessType: string;
  docsUrl: string;
  developerUrl: string;
  scopes: string[];
  environments: string[];
  nextAction: string;
  lastVerified: string;
  /** @nullable */
  validationMessage?: string | null;
  supports?: string[];
}

export interface ValidationResult {
  providerSlug: string;
  status: string;
  message: string;
  checkedAt: string;
}

export interface VaultStatus {
  state: string;
  credentialCount: number;
  environmentCount: number;
  autoLockMinutes: number;
  telemetryEnabled: boolean;
  /** @nullable */
  lastUnlockedAt?: string | null;
}

export interface Workflow {
  id: string;
  providerSlug: string;
  action: string;
  state: string;
  /**
     * @minimum 0
     * @maximum 100
     */
  progress: number;
  checkpoint: string;
  updatedAt: string;
  /** @nullable */
  nextAction?: string | null;
}

export type InventoryItemMetadata = { [key: string]: unknown };

export interface InventoryItem {
  id: string;
  category: string;
  title: string;
  sku: string;
  /** @minimum 0 */
  quantity: number;
  /** @minimum 0 */
  price: number;
  /** @minimum 0 */
  cost: number;
  status: string;
  metadata?: InventoryItemMetadata;
}

export type InventoryInputMetadata = { [key: string]: unknown };

export interface InventoryInput {
  category: string;
  title: string;
  sku: string;
  /** @minimum 0 */
  quantity: number;
  /** @minimum 0 */
  price: number;
  /** @minimum 0 */
  cost: number;
  metadata?: InventoryInputMetadata;
}

export type InventoryUpdateMetadata = { [key: string]: unknown };

export interface InventoryUpdate {
  /** @minimum 0 */
  quantity?: number;
  /** @minimum 0 */
  price?: number;
  status?: string;
  metadata?: InventoryUpdateMetadata;
}

export type ListBossKeyProvidersParams = {
category?: string;
status?: string;
search?: string;
};

export type ListBossKeyInventoryParams = {
category?: string;
};