# lib/api-zod/src/generated/types/bossKeyOverview.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface BossKeyOverview {
  /**
     * @minimum 0
     * @maximum 100
     */
  readinessScore: number;
  /** @minimum 0 */
  missingConnections: number;
  /** @minimum 0 */
  expiringTokens: number;
  /** @minimum 0 */
  failedValidations: number;
  /** @minimum 0 */
  pausedWorkflows: number;
  /** @minimum 0 */
  inventoryConflicts: number;
  /** @minimum 0 */
  securityAlerts: number;
  discoveredAt: string;
}