# lib/api-zod/src/generated/types/discoveryResult.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DiscoveryResult {
  projectName: string;
  repository: string;
  branch: string;
  providersFound: number;
  environmentIssues: number;
  backupCreated: boolean;
  manifestPath: string;
  completedAt: string;
}