# lib/api-zod/src/generated/types/inventoryInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { InventoryInputMetadata } from './inventoryInputMetadata';

export interface InventoryInput {
  category: string;
  title: string;
  sku: string;
  /** @minimum 0 */
  quantity: number;
  /** @minimum 0 */
  price: number;
  /** @minimum 0 */
  cost: number;
  metadata?: InventoryInputMetadata;
}